REF: split EWMRS into an importable pipeline module
Rename the EWMRS entrypoint to pipeline.py, remove its standalone scheduler script, and keep tandem execution focused on staged rendering with concise layer success logging.
